backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.5.2
redis==5.0.7
celery==5.4.0
alembic==1.14.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
resend==2.19.0
